tests: move/rename catalog item-related
