--- spec/features/history_items/browse_spec.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-feature "History items", as: :user do
-  scenario "Filtering history items by search query" do
-    create :history_item, :taxt, taxt: "typo of Forel"
-    create :history_item, :taxt, taxt: "typo of August"
-
-    visit history_items_path
-    i_should_see "typo of Forel"
-    i_should_see "typo of August"
-
-    fill_in "q", with: "Forel"
-    click_button "Search"
-    i_should_see "typo of Forel"
-    i_should_not_see "typo of August"
-
-    fill_in "q", with: "asdasdasd"
-    click_button "Search"
-    i_should_not_see "typo of Forel"
-    i_should_not_see "typo of August"
-  end
-end
